requirements.txt
Adafruit-Blinka
adafruit-blinka-displayio
adafruit-circuitpython-busdevice
adafruit-circuitpython-neopixel
adafruit-circuitpython-pypixelbuf
adafruit-circuitpython-rgb-display
adafruit-circuitpython-ssd1351
Adafruit-DHT
Adafruit-PlatformDetect
Adafruit-PureIO
asn1crypto
binary
certifi
chardet
cryptography
entrypoints
getmac
idna
keyring
keyrings.alt
numpy
Pillow
pip
pycrypto
pyftdi
PyGObject
PyMySQL
pyserial
python-apt
pyusb
pyxdg
PyYAML
recordclass
requests
rpi-ws281x
RPi.GPIO
SecretStorage
setuptools
six
smbus
spidev
ssh-import-id
sysv-ipc
urllib3
wheel
